Update scripts based on Daniel's description of his workflow
Also fixes a bug in process_issues.

+++ bin/build_lists_only_incremental.sh
@@ -3,22 +3,23 @@ if output=$(git status --porcelain) && [ -z "$output" ]; then
     # Working directory clean
     git rev-parse HEAD > mailing/commit.new
 else
-    echo Working directory is dirty! Exiting.
-    echo For incremental builds to reliably work, please commit your changes.
-    exit 1
+    echo Working directory is dirty! Later incremental update will not work.
+    rm -f mailing/commit.new mailing/commit
 fi
 
-if [ -f mailing/commit ]; then 
-   git diff --name-only "$(cat mailing/commit)" > mailing/changed_files.txt
+if [ -f mailing/commit.prev ]; then
+   git diff --name-only "$(cat mailing/commit.prev)" > mailing/changed_files.txt
+   git ls-files -o --exclude-standard xml/ >> mailing/changed_files.txt # add any untracked files in xml/ to the list of changed files
+else
+   rm -f mailing/changed_files.txt
 fi
 
 bin/lists
 
 if [ -f mailing/lwg-active.html ]; then
-    test -f mailing/commit && cp mailing/commit mailing/commit.prev
-    mv mailing/commit.new mailing/commit
+    test -f mailing/commit.new && mv mailing/commit.new mailing/commit
 else
-    rm -f mailing/changed_files.txt mailing/updated_files.txt
+    rm -f mailing/changed_files.txt mailing/updated_files.txt mailing/commit.new
     echo ***********************************
     echo ********** build failure **********
     echo ***********************************
